Update TypeScript types to allow listeners that return Promises
Emittery already supports awaiting event emission, which in turn awaits all the listeners that return a Promise. If a caller awaits `emit`, all the listeners will have resolved by the time the outer promise resolves, which is great!

This adds a test for that behaviour, and updates the TypeScript types to support these async listeners. There's no new or changed behaviour.